fix: update social reporter, add mentorID in discord reporter
